# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/employees:
    get:
      tags:
        - Employees
      summary: List Employees
      description: Get a paginated list of Employees with optional filtering and sorting
      operationId: listEmployees
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Employees
      summary: Create Employee
      description: Create a new Employee
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/employee_identities:
    get:
      tags:
        - EmployeeIdentities
      summary: List EmployeeIdentities
      description: Get a paginated list of EmployeeIdentities with optional filtering and sorting
      operationId: listEmployeeIdentities
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeIdentityList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - EmployeeIdentities
      summary: Create EmployeeIdentity
      description: Create a new EmployeeIdentity
      operationId: createEmployeeIdentity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeIdentityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeIdentity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Employee:
      type: object
      required:
        - id
        - employee_number
        - first_name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_number:
          type: string
        user_id:
          nullable: true
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          nullable: true
          type: string
        email:
          nullable: true
          type: string
        mobile_phone:
          nullable: true
          type: string
        phone:
          nullable: true
          type: string
        birth_place:
          nullable: true
          type: string
        birth_date:
          nullable: true
          type: string
          format: date
        gender:
          nullable: true
          type: string
        marital_status:
          nullable: true
          type: string
        blood_type:
          nullable: true
          type: string
        religion_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateEmployeeRequest:
      type: object
      required:
        - employee_number
        - first_name
        - metadata
      properties:
        employee_number:
          type: string
        user_id:
          nullable: true
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          nullable: true
          type: string
        email:
          nullable: true
          type: string
        mobile_phone:
          nullable: true
          type: string
        phone:
          nullable: true
          type: string
        birth_place:
          nullable: true
          type: string
        birth_date:
          nullable: true
          type: string
          format: date
        gender:
          nullable: true
          type: string
        marital_status:
          nullable: true
          type: string
        blood_type:
          nullable: true
          type: string
        religion_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    EmployeeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    EmployeeIdentity:
      type: object
      required:
        - id
        - employee_id
        - identity_type
        - identity_number
        - is_permanent
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        identity_type:
          type: string
        identity_number:
          type: string
        identity_expiry_date:
          nullable: true
          type: string
          format: date
        is_permanent:
          type: boolean
        metadata:
          type: object
...
    CreateEmployeeIdentityRequest:
      type: object
      required:
        - employee_id
        - identity_type
        - identity_number
        - is_permanent
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        identity_type:
          type: string
        identity_number:
          type: string
        identity_expiry_date:
          nullable: true
          type: string
          format: date
        is_permanent:
          type: boolean
        metadata:
          type: object
...
    EmployeeIdentityList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeIdentity'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []